url_download_request/requirements.txt
requests==2.20.1
rfc3987==1.3.8
PyMySQL==0.9.3
PyYAML==3.13